refactor: split api server endpoints
