+++ lmdeploy/serve/openai/endpoints/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+"""OpenAI and LMDeploy endpoint router assembly."""
+
+from .router import create_openai_router
+
+__all__ = ['create_openai_router']
